+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import nextstep.mvc.adapter.HandlerAdapter;
+import nextstep.mvc.handlermapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
