+++ mvc/src/main/java/nextstep/mvc/controller/ForwardController.java
@@ -1,8 +1,7 @@
-package nextstep.mvc.controller.asis;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Objects;
 
 public class ForwardController implements Controller {
