+++ mvc/src/main/java/nextstep/mvc/handlermapping/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.handlermapping;
 
 import nextstep.web.support.RequestMethod;
